requirements.txt
streamlit
google-generativeai
supabase
pandas
pdfplumber
pytesseract
pdf2image
openpyxl
plotly
tenacity
fastapi
uvicorn[standard]
python-multipart
